+++ src/latex-commands/styling.ts
@@ -850,6 +850,7 @@ defineFunction('mathop', '{:auto}', {
           ? atom.attachLimits(context, { base })
           : atom.attachSupsub(context, { base });
     }
+    if (atom.caret) base.caret = atom.caret;
     return new Box(atom.bind(context, base), {
       type: 'op',
       isSelected: atom.isSelected,