+++ src/addons/static-render.ts
@@ -38,7 +38,11 @@ export type StaticRenderOptionsPrivate = StaticRenderOptions & {
   format?: string;
 };
 
-function findEndOfMath(delimiter, text, startIndex: number): number {
+function findEndOfMath(
+  delimiter: string,
+  text: string,
+  startIndex: number
+): number {
   // Adapted from
   // https://github.com/Khan/perseus/blob/master/src/perseus-markdown.jsx
   let index = startIndex;
@@ -71,13 +75,11 @@ function splitAtDelimiters(
 ): {
   type: string;
   data: string;
-  rawData?: string;
   mathstyle?: string;
 }[] {
   const finalData: {
     type: string;
     data: string;
-    rawData?: string;
     mathstyle?: string;
   }[] = [];
 
@@ -133,12 +135,7 @@ function splitAtDelimiters(
           if (format === 'ascii-math')
             [, formula] = parseMathString(formula, { format: 'ascii-math' });
 
-          finalData.push({
-            type: 'math',
-            data: formula,
-            rawData: text.slice(currIndex, nextIndex + rightDelim.length),
-            mathstyle,
-          });
+          finalData.push({ type: 'math', data: formula, mathstyle });
 
           currIndex = nextIndex + rightDelim.length;
         }
@@ -146,25 +143,23 @@ function splitAtDelimiters(
         lookingForLeft = !lookingForLeft;
       }
 
-      if (currIndex < text.length) {
-        finalData.push({
-          type: 'text',
-          data: text.slice(currIndex),
-        });
-      }
+      if (currIndex < text.length)
+        finalData.push({ type: 'text', data: text.slice(currIndex) });
     } else finalData.push(startDatum);
   }
 
   return finalData;
 }
 
+/** Parse the input text, looking for math delimiters. Return
+ * an array of strings and math strings.
+ */
 function splitWithDelimiters(
   text: string,
   options: StaticRenderOptionsPrivate
 ): {
   type: string;
   data: string;
-  rawData?: string;
   mathstyle?: string;
 }[] {
   let data = [{ type: 'text', data: text }];
@@ -237,9 +232,8 @@ function createMathMLNode(
 function createMarkupNode(
   text: string,
   options: StaticRenderOptionsPrivate,
-  mathstyle: 'displaystyle' | 'textstyle',
-  createNodeOnFailure: boolean
-): HTMLElement | Text | null {
+  mathstyle: 'displaystyle' | 'textstyle'
+): HTMLElement | Text {
   // Create a node for displaying math.
   //   This is slightly ugly because in the case of failure to create the markup,
   //   sometimes a text node is desired and sometimes not.
@@ -261,29 +255,25 @@ function createMarkupNode(
     return element;
   } catch (error: unknown) {
     console.error("Could not parse'" + text + "' with ", error);
-    if (createNodeOnFailure) return document.createTextNode(text);
+    return document.createTextNode(text);
   }
-
-  return null;
 }
 
 function createAccessibleMarkupPair(
   latex: string,
   mathstyle: 'displaystyle' | 'textstyle' | '',
-  options: StaticRenderOptionsPrivate,
-  createNodeOnFailure: boolean
-): Node | null {
+  options: StaticRenderOptionsPrivate
+): Node {
   // Create a math node (a span with an accessible component and a visual component)
   // If there is an error in parsing the latex, 'createNodeOnFailure' controls whether
   //   'null' is returned or an accessible node with the text used.
   const markupNode = createMarkupNode(
     latex,
     options,
-    mathstyle ? mathstyle : 'textstyle',
-    createNodeOnFailure
+    mathstyle ? mathstyle : 'textstyle'
   );
   const accessibleContent = options.renderAccessibleContent ?? '';
-  if (markupNode && /\b(mathml|speakable-text)\b/i.test(accessibleContent)) {
+  if (/\b(mathml|speakable-text)\b/i.test(accessibleContent)) {
     const fragment = document.createElement('span');
     if (/\bmathml\b/i.test(accessibleContent) && options.renderToMathML)
       fragment.append(createMathMLNode(latex, options));
@@ -312,34 +302,37 @@ function scanText(
   text: string,
   options: StaticRenderOptionsPrivate
 ): Node | null {
-  // If the text starts with '\begin'... (this is a MathJAX behavior)
+  if (/^\s*$/.test(text)) return null;
+
+  console.log('scanText', text);
+
   let fragment: Node | null = null;
+
+  // If the text starts with '\begin'... (this is a MathJAX behavior)
   if (options.TeX?.processEnvironments && /^\s*\\begin/.test(text)) {
     fragment = document.createDocumentFragment();
-    const node = createAccessibleMarkupPair(text, '', options, true);
-    if (node) fragment.appendChild(node);
-  } else {
-    if (!text.trim()) return null;
-    const data = splitWithDelimiters(text, options);
-    if (data.length === 1 && data[0].type === 'text') {
-      // This text contains no math. No need to continue processing
-      return null;
-    }
+    fragment.appendChild(createAccessibleMarkupPair(text, '', options));
+    return fragment;
+  }
 
-    fragment = document.createDocumentFragment();
+  const data = splitWithDelimiters(text, options);
 
-    for (const datum of data) {
-      if (datum.type === 'text')
-        fragment.appendChild(document.createTextNode(datum.data));
-      else {
-        const node = createAccessibleMarkupPair(
+  // If the text contains no math, no need to continue processing
+  if (data.length === 1 && data[0].type === 'text') return null;
+
+  fragment = document.createDocumentFragment();
+
+  for (const datum of data) {
+    if (datum.type === 'text')
+      fragment.appendChild(document.createTextNode(datum.data));
+    else {
+      fragment.appendChild(
+        createAccessibleMarkupPair(
           datum.data,
           datum.mathstyle === 'textstyle' ? 'textstyle' : 'displaystyle',
-          options,
-          true
-        );
-        if (node) fragment.appendChild(node);
-      }
+          options
+        )
+      );
     }
   }
 
@@ -351,13 +344,13 @@ function scanElement(
   options: StaticRenderOptionsPrivate
 ): void {
   if (element.childNodes.length === 1 && element.childNodes[0].nodeType === 3) {
+    const text = element.childNodes[0].textContent ?? '';
+
     // This is a node with textual content only. Perhaps an opportunity
     // to simplify and avoid creating extra nested elements...
-    const text = element.childNodes[0].textContent ?? '';
     if (options.TeX?.processEnvironments && /^\s*\\begin/.test(text)) {
       element.textContent = '';
-      const node = createAccessibleMarkupPair(text, '', options, true);
-      if (node) element.append(node);
+      element.append(createAccessibleMarkupPair(text, '', options));
       return;
     }
 
@@ -366,21 +359,19 @@ function scanElement(
       // The entire content is a math expression: we can replace the content
       // with the latex markup without creating additional wrappers.
       element.textContent = '';
-      const node = createAccessibleMarkupPair(
-        data[0].data,
-        data[0].mathstyle === 'textstyle' ? 'textstyle' : 'displaystyle',
-        options,
-        true
+      element.append(
+        createAccessibleMarkupPair(
+          data[0].data,
+          data[0].mathstyle === 'textstyle' ? 'textstyle' : 'displaystyle',
+          options
+        )
       );
-      if (node) element.append(node);
       return;
     }
 
-    if (data.length === 1 && data[0].type === 'text') {
-      // This element only contained text with no math. No need to
-      // do anything.
-      return;
-    }
+    // If this element only contained text with no math, no need to
+    // do anything.
+    if (data.length === 1 && data[0].type === 'text') return;
   }
 
   // Iterate backward, as we will be replacing childNode with a documentfragment
@@ -401,8 +392,6 @@ function scanElement(
         i--;
         content = ((element.childNodes[i] as Text).textContent ?? '') + content;
       }
-      content = content.trim();
-      if (!content) continue;
       const frag = scanText(content, options);
       if (frag) {
         i += frag.childNodes.length - 1;
@@ -416,6 +405,8 @@ function scanElement(
       const tag = childNode.nodeName.toLowerCase();
       if (tag === 'script') {
         const scriptNode = childNode as HTMLScriptElement;
+        if (scriptNode.type === 'module' || scriptNode.type === 'javascript')
+          continue;
         let textContent: string | undefined = undefined;
         if (options.processScriptTypePattern?.test(scriptNode.type))
           textContent = scriptNode.textContent ?? '';
@@ -440,38 +431,16 @@ function scanElement(
               style = value.trim() === 'display' ? 'displaystyle' : 'textstyle';
           }
 
-          const span = createAccessibleMarkupPair(
-            textContent,
-            style,
-            options,
-            true
-          );
-          if (span) scriptNode.parentNode!.replaceChild(span, scriptNode);
+          replaceWithMath(scriptNode, textContent, style, options);
         }
       } else {
         if (options.texClassDisplayPattern?.test(el.className)) {
-          const formula = el.textContent;
-          el.textContent = '';
-          const node = createAccessibleMarkupPair(
-            formula ?? '',
-            'displaystyle',
-            options,
-            true
-          );
-          if (node) el.append(node);
+          replaceWithMath(el, el.textContent ?? '', 'displaystyle', options);
           continue;
         }
 
         if (options.texClassInlinePattern?.test(el.className)) {
-          const formula = el.textContent;
-          el.textContent = '';
-          const node = createAccessibleMarkupPair(
-            formula ?? '',
-            'textstyle',
-            options,
-            true
-          );
-          if (node) element.append(node);
+          replaceWithMath(el, el.textContent ?? '', 'textstyle', options);
           continue;
         }
 
@@ -591,4 +560,16 @@ export function _renderMathInElement(
       );
     }
   }
+}
+
+function replaceWithMath(
+  el: HTMLElement,
+  latex: string,
+  style: 'textstyle' | 'displaystyle',
+  options: StaticRenderOptionsPrivate
+): void {
+  el.parentNode!.replaceChild(
+    createAccessibleMarkupPair(latex, style, options),
+    el
+  );
 }