+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## 0.101.2 _2024-11-15_
 
 ### Issues Resolved